# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpalLuxe Realty API",
        "description": "OpalLuxe Realty API Documentation - Real Estate Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/properties/featured": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get featured properties",
                "description": "Retrieve a list of featured properties (up to 6)",
                "operationId": "d5aa48e0f31bf9a239b8ce80a6c6a94b",
                "responses": {
                    "200": {
                        "description": "Featured properties retrieved successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/properties/best-deals": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get best deal properties",
                "description": "Retrieve a list of best deal properties (up to 6)",
                "operationId": "390d13c1118c987c0c0c48007350a967",
                "responses": {
                    "200": {
                        "description": "Best deals retrieved successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/properties/{slug}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get a single property by slug",
                "description": "Retrieve detailed information about a specific property. View count is automatically incremented.",
                "operationId": "3b9a5917561a9d6e3c3260bfff324ed5",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Property slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property retrieved successfully"
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/properties/{id}": {
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Update a property",
                "description": "Update an existing property (requires authentication)",
                "operationId": "19a6c5948a5a419de858916d8a4ab7c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Property ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "listing_type": {
                                        "type": "string",
                                        "enum": [
                                            "sale",
                                            "rent"
                                        ]
                                    },
                                    "bedrooms": {
                                        "type": "integer"
                                    },
                                    "bathrooms": {
                                        "type": "integer"
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully"
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Delete a property",
                "description": "Delete a property (requires authentication)",
                "operationId": "acabc512dd07a218f8c1a4cf46a922a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Property ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property deleted successfully"
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}